finalize myrmex

+++ src/main/java/com/github/alexthe666/iceandfire/block/BlockMyrmexBiolight.java
@@ -40,7 +40,7 @@ public class BlockMyrmexBiolight extends BushBlock {
         if (!worldIn.isRemote) {
             this.updateState(state, worldIn, pos, state.getBlock());
         }
-        if (!worldIn.getBlockState(pos.up()).isSolid()) {
+        if (!worldIn.getBlockState(pos.up()).isSolid() && worldIn.getBlockState(pos.down()).getBlock() != this) {
             worldIn.destroyBlock(pos, true);
         }
     }
